refactor: util types 분리
